Update to AileronAssist 1.6 and MatrixNav 1.7
git-svn-id: https://gentlenav.googlecode.com/svn/trunk@25 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MatrixNav/optionsRmat.h
@@ -15,15 +15,15 @@
 #define RADPERSEC ((long long)5632.0/SCALEGYRO)
 // one radian per second, in AtoD/2 units
 
-#define GRAVITYM ((long long)980) 
+#define GRAVITYM ((long long)980.0) 
 // 100 times gravity, meters/sec/sec
 
 #define ACCELSCALE ((long) ( GRAVITY/GRAVITYM ) )
 
-#define CENTRISCALE (long) (((long long)519168)*GRAVITY)/(RADPERSEC*GRAVITYM)
+#define CENTRISCALE (long) ((((long long)519168.0)*GRAVITY)/((long long)RADPERSEC*GRAVITYM))
 // scale factor in multiplying omega times velocity to get centrifugal acceleration
 
-#define CENTRIFSAT (long) (GRAVITYM*RADPERSEC)/(GRAVITY*((long long)32))
+#define CENTRIFSAT (long) ((GRAVITYM*RADPERSEC)/(GRAVITY*((long long)32)))
 // saturation limit for the centrifugal adjustment to avoid numeric overflow
 
 #define RMAX   0b0100000000000000	//	1.0 in 2.14 fractional format
